+++ lib/src/res/vs_text.dart
@@ -1,4 +1,7 @@
+import 'dart:nativewrappers/_internal/vm/lib/developer.dart';
+
 import 'package:flutter/services.dart' show rootBundle;
+import 'package:virtusize_flutter_sdk/src/utils/virtusize_constants.dart';
 
 import '../../src/models/virtusize_enums.dart';
 import '../../src/res/vs_font.dart';
@@ -14,11 +17,19 @@ class VSText {
   /// Loads the localization from the local i18n json files and the custom font info
   static Future<VSText> load(String localeName, VSLanguage language) async {
     VirtusizeLocalization localization = VirtusizeLocalization(
-        await rootBundle.loadString('packages/virtusize_flutter_sdk/assets/i18n/$localeName.json'));
+      await rootBundle.loadString(
+        'packages/virtusize_flutter_sdk/assets/i18n/$localeName.json',
+      ),
+    );
     try {
-      language = VSLanguage.values.firstWhere((lang) => lang.langCode == localeName);
-    } catch(e) {
-      print('Could not get the language by the locale name $localeName, $e');
+      language = VSLanguage.values.firstWhere(
+        (lang) => lang.langCode == localeName,
+      );
+    } catch (e) {
+      log(
+        'Could not get the language by the locale name $localeName, $e',
+        name: virtusizeLogLabel,
+      );
     }
     return VSText._(localization, VSFont(language));
   }