+++ lib/src/widgets/virtusize_inpage_mini.dart
@@ -41,6 +41,7 @@ class _VirtusizeInPageMiniState extends State<VirtusizeInPageMini> {
   late final StreamSubscription<VSText> _vsTextSubscription;
   late final StreamSubscription<ProductDataCheck> _pdcSubscription;
   late final StreamSubscription<Recommendation> _recSubscription;
+  late final StreamSubscription<String> _errorSubscription;
 
   VSText _vsText = IVirtusizeSDK.instance.vsText;
   ProductDataCheck? _productDataCheck;
@@ -86,6 +87,18 @@ class _VirtusizeInPageMiniState extends State<VirtusizeInPageMini> {
         _isLoading = false;
       });
     });
+
+    _errorSubscription = IVirtusizeSDK.instance.productErrorStream.listen((
+      externalProductId,
+    ) {
+      if (_productDataCheck?.externalProductId != externalProductId) {
+        return;
+      }
+      setState(() {
+        _isLoading = false;
+        _hasError = true;
+      });
+    });
   }
 
   @override
@@ -93,6 +106,7 @@ class _VirtusizeInPageMiniState extends State<VirtusizeInPageMini> {
     _vsTextSubscription.cancel();
     _pdcSubscription.cancel();
     _recSubscription.cancel();
+    _errorSubscription.cancel();
     super.dispose();
   }